+++ lib/flutter_homescreen_widget_platform_interface.dart
@@ -3,27 +3,27 @@ import 'dart:typed_data';
 
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-import 'flutter_widget_kit_method_channel.dart';
+import 'flutter_homescreen_widget_method_channel.dart';
 
-/// The platform interface contract for [FlutterWidgetKit].
+/// The platform interface contract for [FlutterHomescreenWidget].
 ///
 /// Platform implementations must extend this class and set [instance] to
 /// themselves before any plugin methods are called.
-abstract class FlutterWidgetKitPlatform extends PlatformInterface {
-  /// Constructs a [FlutterWidgetKitPlatform].
-  FlutterWidgetKitPlatform() : super(token: _token);
+abstract class FlutterHomescreenWidgetPlatform extends PlatformInterface {
+  /// Constructs a [FlutterHomescreenWidgetPlatform].
+  FlutterHomescreenWidgetPlatform() : super(token: _token);
 
   static final Object _token = Object();
 
-  static FlutterWidgetKitPlatform _instance = MethodChannelFlutterWidgetKit();
+  static FlutterHomescreenWidgetPlatform _instance = MethodChannelFlutterHomescreenWidget();
 
-  /// The current default [FlutterWidgetKitPlatform] instance.
+  /// The current default [FlutterHomescreenWidgetPlatform] instance.
   ///
-  /// Defaults to [MethodChannelFlutterWidgetKit].
-  static FlutterWidgetKitPlatform get instance => _instance;
+  /// Defaults to [MethodChannelFlutterHomescreenWidget].
+  static FlutterHomescreenWidgetPlatform get instance => _instance;
 
-  /// Sets the default [FlutterWidgetKitPlatform] instance.
-  static set instance(FlutterWidgetKitPlatform instance) {
+  /// Sets the default [FlutterHomescreenWidgetPlatform] instance.
+  static set instance(FlutterHomescreenWidgetPlatform instance) {
     PlatformInterface.verifyToken(instance, _token);
     _instance = instance;
   }